+++ docs/changelog.rst
@@ -12,6 +12,7 @@ Changelog
 
 * HTML5: fixed GL_INVALID_OPERATION (or missing geometry) when rendering untextured meshes.
 * Data Compiler: fixed shaders not including code in some circumnstances.
+* Lua API: added an optional exit_code parameter to ``Device.quit()``.
 
 0.55.0 --- 03 Apr 2025
 ----------------------